Fix link to EncryptionMaterialsProvider interface
The current link to the EncryptionMaterialsProvider interface is to the
S3 package. However, EMR is expecting it to be in the EMR package, so
updating the link to the documentation for the interface in the EMR SDK.
Additionally, git indicates this was only available in the SDK starting
in 1.11.10 (specifically introduced in
aws/aws-sdk-java@cafdbe14587d6d0d1d532d67906617d5adc8c721), so updating
the version reference.